Make the project a bit better organized, in order to make it callable from other places as well
