+++ example/01-basic/how-to-run.md
@@ -1,0 +1,1 @@
+This is a simple example of running the active matter FDM. To run, execute `python example.py`
